Implement threading in bolo2rrd
The bolo2rrd subscriber is now made up of a creator thread and multiple
updater threads.  The creator thread is responsible for provisioning new
round-robin database files, and the updater threads handle updates to
existing RRDs.  In this way, the addition of new monitoring (an
I/O-costly proposition) does not adversely affect (i.e. block or delay)
any existing monitoring.