Correct submit_* helper functions / update PROTO
The listener-side of the protocol always starts the PDU off with a
timestamp frame.  This was mistakenly removed from the PROTO doc some
time ago, and caused some bugs in the new submit_* helper functions.

This commit also introduces new *_pdu() methods that do what submit_*
did, but allow the caller to send the resulting PDU however they see
fit.