dbolo: fix scheduling zombies bug
If a child process is still running, we should never check the schedule
and attempt to execute it again.  Doing so overwrites the PID for the
previous run, causing Zombie processes to stack up since dbolo will
never be able to wait() for that original process.  Since dbolo doesn't
exit, those processes never become children of init(1), so they tend to
stick around.