Migrate ctap test suite to shell
It makes more sense, and ensures that we are (A) testing what actually
gets used by the end users, (B) understand how the interfaces are used,
and (C) find and remove dead code.

All of the old t/*.c files are gone, the dependency on libctap is
lifted, and we gain a ton of flexibility in our test suites.  Win!