Prefer network order functions from libvigor (#90)
With the recent addition of 64bit network order functions to the most
recent version of libvigor, binf.c no longer compiles because it is
redefining these same functions. While this could be resolved with
defining these functions before including the libvigor header file, I
believe this solution is slightly quicker and allows for the depracation
of the static functions.